Optimized foo. foo/ - path behavior for terminals
This builds on the commit introducing the refinement optimization
for /foo and .foo by bringing in the optimization of terminal forms
like foo. and foo/

It also introduces some fledgling support for the proposed behavior
that terminal dots prohibit function fetches or execution, while
terminal slashes enforce function fetches and execution.  It's very
kludgey and is just patched in to get a feel for it prior to a more
general rewrite of the pathing mechanisms.