Use AS BLOCK! to get FOR-EACH on all paths
To quickly allow FOR-EACH to work on paths and avoid code repetition,
this simply uses AS to convert paths to blocks (if they aren't
already block representations) when being iterated by FOR-EACH.  A
more optimized implementation should only be undertaken when there's
a better model laid out for how generic iteration works (e.g. over
user-defined types, and what the invariants are e.g. in suspension
of stackless frames that are in mid-enumeration).