Replace dot-bearing names in source-analysis
The source analysis code made use of identifiers with dots in them.
Clever tricks might be attempted to preserve this and use new
mechanisms for namespacing (into objects, or something more fancy).
However, the bootstrap dependency of the C lexical code makes that
impractical, so this just uses hyphens instead.