Unification of ISSUE! and CHAR! Datatypes
This rethinks ISSUE! as an immutable stringlike datatype which has no
index position.  That redefinition allows it to be implemented as a
series of encoded UTF-8 bytes directly in a cell (if the encoded content
is short enough).

With this ability, a single-character issue can take the place of CHAR!.
Thus CHAR! becomes a "type constraint" for whenever an issue contains
0 or 1 elements...with the 0 case of an empty issue (`#`) representing
the NUL codepoint.

TO conversions from INTEGER! <=> ISSUE! are temporarily disallowed.
Long term, it is expected that:

    >> to integer! #1
    == 1

    >> to issue! 1
    == #1

But this means there needs to be a different way to translate to
codepoints.  So this adds CODEPOINT OF as the method for extracting a
codepoint from a single-length ISSUE!.  MAKE CHAR! is preserved for
the moment as one way of getting an ISSUE! from an INTEGER! codepoint,
but AS ISSUE! of an INTEGER! is revealed as another option.

The code focuses on accomplishing the storage efficiency goals, but
has not been vetted for CPU efficiency on the usage of these new
character types...and there's significant debug code to make sure it
is working.  But long term, it should be more efficient than when
CHAR! and ISSUE! were separate.