Fix Analyze_String_For_Scan with cell-embedded UTF-8
Analyze_String_For_Scan() fabricated a temporary value with a changed
index in order to leverage existing VAL_STRING_AT() functions.  With
UTF-8 payloads living in issues, this would mean pointing at the
fabricated cell...which would not be any good after returning up
the stack (the way a pointer into a series would be).

That was bad.  So this rewrites Analyze_String_For_Scan() to not use
that fabricated local, and operate in a more conventional way.  Doing
such a fix also involved giving an accurate size to the scanner (there
had been a bug which was passing trailing whitespace), exposing the
missing `limit` feature in the scanner.

This changes the scanner to use an unlimited form and crop the size
of a word after the token search.  To make usages of "UNKNOWN" clearer,
it is changed to UNLIMITED...as all the UNKNOWNs were passed as a
limit parameter.