:X :Y/Y :(Z) subvert quoting, not :X :Y/Y (Z)
GET-WORD! and GET-PATH! would subvert "soft quoting" in Rebol2:

    rebol2>> foo: func ['a] [print mold a]

    rebol2>> x: 10

    rebol2>> foo x
    x

    rebol2>> foo :x
    10

However, groups (PAREN!) did not subvert the quoting:

    rebol2>> foo (1 + 2)
    (1 + 2)

This was changed in R3-Alpha in order to allow more complex calculations
to escape quoted slots...at the cost of requiring a slot to be "hard
quoted" in order to get at an actual GROUP!:

    r3-alpha>> foo: func ['a] [print mold a]

    r3-alpha>> foo (1 + 2)
    3

But Ren-C offers a new datatype in the GET-GROUP!.  This means that
"normal-looking" groups can be "quoted normally", while colon-prefixed
groups can be escaped:

    >> foo: func ['a] [print mold a]

    >> foo (1 + 2)
    (1 + 2)

    >> foo :(1 + 2)
    3

This makes the quoting subversion have more of a consistency with the
GET-XXX! types.  And having it possible to have plain GROUP!s be
quoted in a partially evaluative slot opens the doors to making
branching constructs not evaluate plain GROUP!s unless the branch runs:

    >> branchy: func [n] [pick [[<a>] [<b>]] n]

    >> either true (print "a" branchy 1) (print "b" branchy 2)
    a
    == <a>

Those wishing the historical behavior where a GROUP! would run at the
callsite regardless of the branching can still get it, using escaping:

    >> either true :(print "a" branchy 1) :(print "b" branchy 2)
    a
    b
    == <a>

With the behavior for GET-GROUP! understood for branching, this opens
the door for defining what the SYM-GROUP! variations would do.  So
they now allow a branch to request their results "as-is" with no
voidification:

    >> type of (if true [null])
    == #[datatype! void!]

    >> type of (if true @[null])
    ; null