COLLECT with no KEEPs returns empty block, not NULL
For a short period of time, the concept was tried that COLLECT with
no keeps would be NULL.

    >> collect []
    ; null

One idea to make this palatable was that if you TRY a NULL, you would
get a BLANK!, which would act equivalently to empty blocks for many
functions (e.g. FOR-EACH would accept blanks and just not do anything,
as it would for an empty block).

Another aspect was that you could simply KEEP an empty block somewhere
inside the COLLECT and that would count as "some material":

    >> collect [keep [], if false [keep <whatever>]]
    == []

At the time, this was thought to offer an interesting piece of data
from the outside of the collect...to be able to test whether it had
collected anything or not without using EMPTY?.  It offered performance
advantages by not needing to allocate blocks when nothing was collected
as well.

It didn't really wind up being as much of a win as it seemed.  So the
NULL-if-no-KEEP version was made a core routine called COLLECT* and
then COLLECT was built on top of that.

However, R3C was apparently picked right before it was changed back.
This goes ahead and patches it.