Get the NMAKE version of Rebmake working again
With the plan being to pare down rebmake significantly, a good first
step in that direction is to migrate away from the generation of large
XML Visual Studio files.

NMAKE support was in the initial Rebmake, but never used.  Hence it
had a couple of aspects that weren't working.  This patches it to
get nmake producing an executable.