Switch binding to use full node instead of misc
The "Binder" abstraction was introduced to make code that mapped
words to integers more abstract.  To try and paint a picture that
a word could point to multiple bindings, it treated the MISC()
field of word symbols as a bitfield struct with two indices.

All along the idea was that this field might "pop out" into a linked
list, but be able to efficiently represent just an integer index
in the average case.  Currently with the "sea of words" concept, the
working idea is that permanent entries that actually represent the
variables for contexts themselves (such as lib) would be linked to,
and then other bindings could temporarily be "popped in" to be
higher priority.

Node allocation from memory pools has to be very fast, so this uses
unmanaged nodes during the binding process to store an arbitrary
mapping element for the bind.

This is a first step on the path to making the "sea of words" have
a place to link in pointers for lib and user context declarations.