Use interning ability on Base/Sys/Mezz
This uses the interning that is done at scanner time to make the
definitions in LIB and SYS able to see new definitions that appear in
LIB and SYS later.

It also simplifies the Startup_Mezzanine() routine so that it uses
the libRebol API forms instead of lower-level calls, which makes
what's going on clearer.