Allow IMPORT on non-modules
This permits the importation of things that are not marked by Type:
of 'Module in their headers.

This gives compatibility with the bootstrap executable on scripts
which need to not run isolated, due to their desire to run with the
bootstrap shim definitions.  Making them into modules leads to
isolation and they use the bootstrap executable's lib.