Change tests to modules/import for DO isolation
With "DO isolation", you can't get imports from a plain script
(unless it is stylized to return an object as the result of the
DO, but then if you have more than one export you will have to
access them with member notation).

Hence the old method of using DO to run the scripts in the
tests no longer works.  This changes to using modules and IMPORT.