Remove module body scan for EXPORT and HIDDEN
Previously the body of modules was searched for the word EXPORT
literally by the MODULE command.

This doesn't allow conditional exportation or other constructions of
the list of exported things at runtime.  It also changes the source
code, which can be confusing (figuring out this code for doing the
scan was overriding a function named EXPORT was difficult).

This switches to reliance on the definitional EXPORT which is added
to the module as a specialization of SYS/EXPORT*.