Kill redundant module vars
The first attempt at implementing the "sea of words" concept for
modules stored a linked list of indexes to variables declared in a
module.  Those stubs could have stored the variables themselves, but
so much of the context mechanics relied on assuming that the context
was a parallel list of keys and variables that only a small part of
the lookup was ready to use the indices.

This switches over to where the stubs actually store the data for
the variable.  An interim solution for enumeration of modules is
established, which captures a list of the keys as a block and
enumerates over that.  (The technical difficulty is that because
the actual data is interwoven with the global hash table for symbols,
rearrangements of that hash table can happen at any time...and there
is no guarantee on what impact this would have on in-progress
enumerations.)

Better approaches can be considered once more experience with the
model has been gained.