Prelude to WHILE as arity-1, WHILE => LOOP
PARSE's WHILE construct emerged as a much better name than ANY for a
construct that repeats a rule and allows zero matches of the rule
to count as success.

The previous distinction between ANY and WHILE (namely, requiring the
rule to advance the input in order to not break out of the repetition)
was handled by creating a new rule called FURTHER, which could be used
in combination with non-looping rules.

With WHILE in PARSE as arity-1, it creates issues with readability
where things like:

     parse ... [while rule [some "a"]]

There had been a longstanding desire to establish parity between WHILE
and UNTIL, so the sensible choice seems to be to make WHILE arity-1
for consistency.  Then the LOOP function is taken for looping based on
something being true.

    loop [condition] [print "Runs while condition is truthy"]
    loop (func [f] [...]) [print "Runs while function returns truthy"]

Eventually an arity-1 WHILE would be added to match UNTIL.  But that
is left out for the moment.