Introduce necessary REPEATs in Encap
Encapping code on Windows was still using the idea of integers in PARSE
where they could become "argument-taking rules" effectively, obfuscated
by the variable name.

That is not legal anymore, so this forces those to use REPEAT.  At the
same time it improves some of the commenting, changes from using the
more lax `null` for initializations where not needed to ~, and switches
GEN-RULE to take the name quoted to make whatever it is doing a bit
more clear.