Eliminate TO-VALUE synonym for TRY
The original name of the function that would convert NULL to BLANK!
(when NULL was indiciating unset values) was TO-VALUE.  This was
based on a proposal by BrianH that was unknowingly re-suggested by
@HostileFork under the same name...making it seem like a good choice.

After more consideration, this was aliased under the name TRY.  The
TO-VALUE was kept around as a synonym, under the belief that some
circumstances might be "clearer" if expressed as TO-VALUE instead.

However, TRY's "I convert nulls to blanks and pass through everything
else" meaning has become so completely pervasive that having a synonym
really just makes people wonder what TO-VALUE means.  With TRY being
solid, the second definition is just confusing.