Increase coherence of SET? and UNSET? definitions
When NULL was first introduced, it tried to unify the concept of a
variable that was bound but not yet assigned with the notion of the
absence of a variable.  Hence NULL variables were the "true" unset.

Over time, the desirability to use NULL safely made it seem like the
ornery state desired for "unset" variables wasn't a good match.  So
unsets returned to being very much like the UNSET! of Rebol2, but
under a different name...since it was emphasized that *variables*
could be unset, but not values...and that NULL was the only true way
to say something was "unset".

This further evolved to where a complete design for a BAD-WORD! type
emerged, and it seemed acceptable to say that an isotope of the
BAD-WORD! for ~unset~ made a good capture of the intent of a variable
to be "ornery".  So terminology backpedaled a bit to be willing to
say that UNSET? and SET? could be tests *of a variable* for the isotope
state of the ~unset~ word.

So this commit goes ahead and migrates toward that, eliminating the
UNDEFINED? and DEFINED? tests that had been used previously...and
adapting to other ways of testing for NULL.