Replace "literalize" with "meta quote", fix comments
The naming for what are now known as META-XXX! types, the symbol, and
their behavior went through significant changes.  For a time the
process was known as literalization and they were LIT-XXX! types with
an @ symbol, but they've now settled sensibly on caret for the symbol
and META for the name.

This goes through and corrects some lingering uses of the "literalize"
term, and also some comments that were found along the way pertaining
to isotopes that were backwards.  (Originally it was the "friendly"
form that were called "isotopes").