~ usage in bootstrap-shim.r for rebmake robustness
There is a current strategy of creating three levels of "nothingness"
that are easy to access via a single character:

    _ makes the friendly blank, which does not error on word access
    and will also silently "opt-out" of many operations (like APPEND)
    frequently giving back NULL for other queries

    ' makes the slightly less friendly null, which though it does not
    error on word access it will error on most operations, besides a
    few that are specifically designed to react to null

    ~ makes the least friendly ~unset~ isotope, which will error on
    word access and can only be accepted by functions which use the
    ^META parameter class

The bootstrap executable did not distinguish null and unsets, so both
would error on word access.  This means the "slightly less friendly"
option is not available.

However, this creates a bootstrap option for ~ producing null.  That
means the the least friendly option is available for the make process
in a way that roughly matches the current state.

This is leveraged in order to make the rather complex rebmake code
more robust, by pointing out fields that must be assigned vs. those
that are optional.  (Most of those optional fields that are blank would
actually be null if not subject to the bootstrap constraints.)