Use `rd` alias for `rmdir` on windows
If you allow git for windows to put the unix command line tools in
your path, then `rmdir` will be a version that does not understand
"/S" or "/Q".

Generally speaking putting those tools in the path isn't a good idea
if you're going to run any windows-based scripting.  But as the only
speedbump encountered for rebmake was this one, this goes ahead and
changes that to `rd` which is an alias that is not redefined.

Additionally, this adds an existence test so that the commands will
not error when doing `nmake clean` if the build products or object
directories don't exist.