Attempt to patch bootstrap executable for GitHub
When GitHub Actions were first being used, the VM would not run the
bootstrap built executables that existed.

To be expedient and make progress on writing the actions, a current
version of the executable was used instead.  However, this created a
drift to needing to keep supporting that version...even as more
evolutions of the system continued to happen.

As that cost grew, it became not worth it.  This moves toward just
getting the bootstrap executable patched to run on GitHub.