Remove UNLESS usage from bootstrap, exec-file fix
The attempt to use "unsets" as a more ornery initialization for values
that needed to be initialized assumed "exec-file" was a field that
rebmake would require.  However, it was used as the input to
FILE-TO-LOCAL with the option of blank, and if it was not present it
was used as the right hand side of an UNLESS.

UNLESS design is still somewhat complex and tied into the philosophy
of putting GROUP!s in executable suspension, as in `(...) and (...)`
where short-circuit evaluation is desired without having to write
`(...) and [...]`.

So this addressed both issues by replacing UNLESS usages in bootstrap
with ANY, and putting exec-file back to BLANK! as a default.