Optimize bootstrap by avoiding shim use in bootstrap shim
This generally avoids use of the bootstrap shim versions of things
(like APPEND, COLLECT) inside of the shim.  Instead, the LIB versions
are preferred.  Hence the wrapping code doesn't run during the shim.

Some exceptions are made for simple transformations like LOOP, REPEAT,
etc.  These are separated to the top of the file.

While this may seem to make the shim a bit more verbose and complex,
writing the shim is already a bit of a black art...and this helps
keep things more consistent.  Future methods of abstracting binding
via sea of words should help things like the shim or Redbol be able
to avoid the LIB/XXX designators, and remain efficient.