No /ONLY use in APPEND/INSERT/CHANGE for bootstrap
The bootstrap executable can actually have a word named ^, and it also
tolerates `^(...)` as being the same as `^ (...)`

While it does not support generic quoting, it can use ^ as an
instruction to put things into a block, and hence achieve somewhat
compatible callsites with the new META-XXX! conventions for the set
of answers to eliminating /ONLY.  If it receives a VOID! value it will
promote it to the WORD! ~bad-word~, so that it undergoes a similar
"loss of orneryness but still looking mean" to how ^ works.  NULL in
also results in NULL out.

This would ultimately need to be applied to FIND/SELECT and other
functions that plan to be in the same family.  For now this just
handles APPEND, INSERT, CHANGE, and the KEEP inside of COLLECT.