Remove ~null~ isotope evaluator aberration
To avoid having to raise an error when NULL pointers were spliced into
API streams, the `~null~` isotope was used.  This "friendly" option
avoids breaking the rules, while still throwing in something visually
obvious...and that evaluates to something that is like pure NULL for
most practical purposes.

However, there was a bit of fear regarding:

    REBVAL *something = nullptr;
    rebValue(something, "else [print {This would print}]");

The problem is that ~null~ turns into a ~null~ isotope, not an actual
pure NULL, so it will run things like ELSE.

This could be solved with:

    REBVAL *something = nullptr;
    rebValue(rebQ(something), "else [print {This would print}]");

However, the thought was that it could instead be changed so that
~null~ would evaluate to NULL, and ' would evaluate to the null isotope.

It was clearly weird, and as the example shows above it doesn't really
solve anything but pushes the burden onto the cases where people are
not taking advantage of the "hack" that turns `null` into something
reified in blocks.  It should be users of the hack--not users of
the rebQ() formal method--who pay for the side effects of the
fabricated ~null~.