Report void isotopes in console
The evaluative results of ^() and ^(comment "hi") are distinct from
the result of ^(~void~); e.g. the meta of an absence produces a ~void~
isotope instead of a ~void~.

    >> ^(first [~void~])
    == '~void~

    >> ^(~void~)
    == ~void~

    >> ^()
    == ~void~  ; isotope

The console implementation used the ^ operator to make the distinction
for most BAD-WORD!s and their isotopes, but conflated the second two
cases.  It then made void isotopes invisible in the console UI.  This
was so that hitting just enter or saying `>> comment "hi"` would appear
to give a complete absence of result.

As understanding the differences between ~void~ isotopes and the
complete absence of an evaluation becomes more important, this does not
help with intuition.  So this makes ~void~ isotopes display, and makes
a special value case for no evaluative product.