ALL/ANY vanish ~void~ isotopes, ALL [] ~void~ when empty
This presents a new philosophy, summarized by:

    >> do [comment "hi"]
    == ~void~  ; isotope

    >> any [do [comment "hi"], 10, 20]
    == 10

    >> all [do [comment "hi"], 10, 20]
    == 20

This joins the idea that `do []` and `do [comment "hi"]` tries to save
the void intent by returning a void isotope.  And incorporates the
idea that isotope voids seen during the operation are seen as being
the same as a pure invisible would be.

It also establishes the baseline for answers to all [] and any []:

    >> all []
    == ~void~  ; isotope

    >> any []
    ; null

This allows a relatively safe way of handling chains of invisibility,
as opposed to erroring on the ~void~ which is not conventionally
either true nor false...but in a piped scenario it does more good than
harm to permit it.