Swap ~void~ isotope in for loops whose body never runs
With functions like ANY and ALL ignoring ~void~ isotopes, the stakes
are raised for there being a good agnostic response for loops that
do not run their bodies to return an ornery value that is neither
true nor false.

Previously they were returning null isotopes so that THEN would run
and ELSE would not (reserving the "failure" case for BREAK).  But
~void~ isotopes now look to be the clear winner for the answer for
LOOP, REPEAT, FOR-EACH, etc.