Fix lib cloning in HELP
The original HELP function used a multiple inheritance concept to
try and merge together the lib and user contexts for searching as
a set.  But while inheritance in R3-Alpha was not thought out, the
multiple inheritance feature was even less so.  So it was removed.

To try and keep "whatever the HELP function was doing" working, it
would manually create a class by copying LIB and appending any variables
in user that were not unset.  However, when the concepts of unsetness
and nullness diverged, it had a problem in proxying values that were
either isotopes or nulls.

This was addressed in a forum post, and the mitigation suggested in
that post is applied here using a META-WORD!:

https://forum.rebol.info/t/1171/3