Delete MATCHED (and MATCHES) for now
MATCHED was an interesting idea about postfix asserting a type, e.g.

    >> 10 + 10 matched integer!
    == 20

    >> 10 + 10 matched text!
    ** Error

While an interesting idea, some changes to MATCH mechanics made the
parameter reversal a bit of an issue, and the parameter conventions
for MATCH need to be improved.  Rather than fix this somewhat esoteric
feature for now, it is just removed.

MATCHES seems like it was a synonym for MATCH?, where the latter more
obviously returns a LOGIC!, so it is also deleted.