Fix TUPLE!/PATH! TO Conversion Hack
It's not entirely clear how much in string space is supposed to work
for TO PATH! and TO TUPLE!.  Historical Rebol has been tolerant, e.g.
saying a string with "a b c" loads as a PATH! of `a/b/c`, but so does
the string of "a/b/c".

To get it working the LOAD mechanics had been reused in the initial
version of a hack.  However, LOAD was changed to be uniform and always
return a BLOCK! of loaded items instead of having the variance that
this trick took advantage of.

This redoes the hack to PARSE the loaded block, affording a general
mechanism for interpreting patterns as sequences.