Remove COMPOSE tests with decorating strings
COMPOSE is willing to put sigils onto WORD!s:

    >> compose [('x):]
    == [x:]

    >> compose [@(first [y z])]
    == @y

There was an idea that if you used a TEXT! this could implicitly
convert strings to unbound words, and then annotate them.

The implementation was removed, at least for the time being, but not
the corresponding tests.  This removes the tests and makes a note
about the feature potentially coming back at some point.