Fix case-sensitivity issue in specialization test
A complex-looking test was skipped over for a while during some churn
in the code.

It turns out the problem with the test was merely an issue with case
sensitivity.

There hasn't been a firm decision on how the language will deal with
case sensitivity.  The current concept is that it might be something
that is left up to the choice of particular projects and dialects...
because there are some very compelling cases where case-sensitivity
can be helpful in expanding the space of identifiers; more than the
"human-friendly" advantages of case-insensitivity.

This fixes the test issue, as part of the goal to shore up the tests
back to the point of zero failures.