Fix SHELL dialect test on Windows
The very speculative SHELL dialect was designed for Linux, not Windows,
and it used the MAKE-FILE! facility on PATH!s and TUPLE!s in order to
do the handling of blocks and GROUP!s inside of them.

But since it produced files, the results would be turned into having
backslashes on Windows.  This feature turns out to not be desirable
as a default, because Windows shells use slashed things for command
line switches etc.

While SHELL is not a current priority, having failing tests on Windows
is distracting, so this standardizes Windows to not flip the direction
of the slashes.