Change TO PATH! test and fix comparison
While very much is still up in the air about TO conversions to paths,
there was a failing test pertaining to an old expectation that

    >> to path! '(a b c)
    == a/b/c

The concept that GROUP!s are more "atomic" than blocks emerged for
various reasons, and so this was instead giving:

    >> to path! '(a b c)
    == /(a b c)

While there's a lot of debate that can be done on this topic, the goal
at the moment is to fix the atrophied failing tests.  So this brings
it into sync with the current behavior, and fixes a comparison needed
by the tests for this representation.