Fix LINES-EXCEEDING in %text-lines.reb
A bad edit in March led to the LINES-EXCEEDING in text-lines.reb not
working.  Due to having a tremendous number of broken tests due to
experimental features, the non-essential feature was overlooked.

Bringing tests back up to date, the fix was relatively simple: to
fix a variable name and move some LET definitions up in the function
so they would be visible to a parse rule.